refactor: Price 리팩터링 - menu 와 product 에서 공통의 Price 객체 사용하도록 수정 - Price Comparable 구현 - Price 에 isGreaterThan 메서드 추가
