+++ src/main/java/kitchenpos/products/tobe/domain/TobeProduct.java
@@ -1,6 +1,7 @@
 package kitchenpos.products.tobe.domain;
 
 import jakarta.persistence.*;
+import kitchenpos.shared.domain.Price;
 import kitchenpos.shared.domain.Profanities;
 
 import java.util.UUID;
