+++ src/main/java/kitchenpos/shared/domain/Price.java
@@ -1,12 +1,13 @@
-package kitchenpos.menus.tobe.domain.menu;
+package kitchenpos.shared.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Objects;
 
 @Embeddable
-public class Price {
+public class Price implements Comparable<Price> {
 
     @Column(name = "price", nullable = false)
     private int price;
@@ -14,14 +15,18 @@ public class Price {
     protected Price() {
     }
 
-    public Price(int price) {
-        if (price < 0) {
-            throw new IllegalArgumentException("가격은 0원 이상이어야 한다.");
-        }
+    private Price(int price) {
         this.price = price;
     }
 
-    int getPrice() {
+    public static Price of(int price) {
+        if (price < 0) {
+            throw new IllegalArgumentException("상품의 가격은 0원 이상이어야 한다.");
+        }
+        return new Price(price);
+    }
+
+    public int getPrice() {
         return price;
     }
 
@@ -36,5 +41,14 @@ public class Price {
     @Override
     public int hashCode() {
         return Objects.hashCode(price);
+    }
+
+    @Override
+    public int compareTo(@NotNull Price other) {
+        return this.price - other.price;
+    }
+
+    public boolean isGreaterThan(int price) {
+        return this.price - price > 0;
     }
 }
