+++ src/test/java/kitchenpos/products/tobe/domain/TobeProductTest.java
@@ -1,6 +1,7 @@
 package kitchenpos.products.tobe.domain;
 
 import kitchenpos.share.domain.FakeProfanities;
+import kitchenpos.shared.domain.Price;
 import kitchenpos.shared.domain.Profanities;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
